requirements.txt
torch
numpy
pandas
scipy
transformers==4.46.3
tokenizers==0.20.3
safetensors==0.4.3
bitsandbytes==0.44.1
nltk
rouge
datasets==3.1.0
peft==0.4.0
accelerate==1.1.1
trl==0.12.1
ipywidgets==7.7.1
huggingface_hub
python-dotenv

